% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_calibration.R
\name{abc_calibration}
\alias{abc_calibration}
\title{Calibrates the reproductive rate and dispersal scales of the pops model.}
\usage{
abc_calibration(
  infected_years_file,
  number_of_observations,
  prior_number_of_observations,
  prior_means,
  prior_cov_matrix,
  params_to_estimate = c(T, T, T, T, F, F),
  number_of_generations = 7,
  generation_size = 1000,
  checks = c(500, 5e+05, 1e+05, 1000),
  infected_file,
  host_file,
  total_plants_file,
  temp = FALSE,
  temperature_coefficient_file = "",
  precip = FALSE,
  precipitation_coefficient_file = "",
  model_type = "SI",
  latency_period = 0,
  time_step = "month",
  season_month_start = 1,
  season_month_end = 12,
  start_date = "2008-01-01",
  end_date = "2008-12-31",
  use_lethal_temperature = FALSE,
  temperature_file = "",
  lethal_temperature = -12.87,
  lethal_temperature_month = 1,
  mortality_on = FALSE,
  mortality_rate = 0,
  mortality_time_lag = 0,
  management = FALSE,
  treatment_dates = c("2008-01-01"),
  treatments_file = "",
  treatment_method = "ratio",
  natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy",
  natural_dir = "NONE",
  natural_kappa = 0,
  anthropogenic_dir = "NONE",
  anthropogenic_kappa = 0,
  pesticide_duration = c(0),
  pesticide_efficacy = 1,
  mask = NULL,
  success_metric = "number of locations and total distance",
  output_frequency = "year",
  output_frequency_n = 1,
  movements_file = "",
  use_movements = FALSE,
  start_exposed = FALSE,
  generate_stochasticity = TRUE,
  establishment_stochasticity = TRUE,
  movement_stochasticity = TRUE,
  deterministic = FALSE,
  establishment_probability = 0.5,
  dispersal_percentage = 0.99,
  quarantine_areas_file = "",
  use_quarantine = FALSE
)
}
\arguments{
\item{infected_years_file}{Raster file with years of initial infection/infestation as individual locations of a pest or pathogen}

\item{number_of_observations}{the number of observations used for this calibartion}

\item{prior_number_of_observations}{the number of total observations from previous calibrations used to weight the posterior distributions (if this is a new calibration this value takes the form of a prior weight (0 - 1))}

\item{prior_means}{A vector of the means of your parameters you are estimating in order from (reproductive_rate, natural_dispersal_distance, percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa, and anthropogenic kappa)}

\item{prior_cov_matrix}{A covariance matrix from the previous years posterior parameter estimation ordered from (reproductive_rate, natural_dispersal_distance, percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa, and anthropogenic kappa)}

\item{params_to_estimate}{A list of booleans specificing which parameters to estimate ordered from (reproductive_rate, natural_dispersal_distance, percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa, and anthropogenic kappa)}

\item{number_of_generations}{the number of generations to use to decrease the uncertainty in the parameter estimation (too many and it will take a long time, too few and your parameter sets will be too wide)}

\item{generation_size}{how many accepted parameter sets should occur in each generation}

\item{checks}{A list of the 4 starting check values in order of # of locations, total min distance, residual error, and # infected. default is (500,500000, 100000, 1000). Starting check values can play a role in speed of calibration and in success of calibration.}

\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster files with number of hosts and standard deviation on those estimates can be based in 3 formats (a single file with number of hosts, a single file with 2 layers number of hosts and standard deviation, or two files 1 with number of hosts and the other with standard deviation of those estimates)}

\item{total_plants_file}{path to raster file with number of total plants}

\item{temp}{boolean that allows the use of temperature coefficients to modify spread (TRUE or FALSE)}

\item{temperature_coefficient_file}{path to raster file with temperature coefficient data for the timestep and number of years specified}

\item{precip}{boolean that allows the use of precipitation coefficients to modify spread (TRUE or FALSE)}

\item{precipitation_coefficient_file}{path to raster file with precipitation coefficient data for the timestep and number of years specified}

\item{model_type}{What type of model most represents your sysetm. Options are "SEI" (Susceptible - Exposed - Infected/Infested) or "SI" (Susceptible - Infected/Infested). Default value is "SI".}

\item{latency_period}{How many times steps does it take to for exposed populations become infected/infested. This is an integer value and must be greater than 0 if model type is SEI.}

\item{time_step}{how often should spread occur options: ('day', 'week', 'month')}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen (integer value between 1 and 12)}

\item{season_month_end}{when does spread end during the year for your pest or pathogen (integer value between 1 and 12)}

\item{start_date}{date to start the simulation with format ('YYYY_MM_DD')}

\item{end_date}{date to end the simulation with format ('YYYY_MM_DD')}

\item{use_lethal_temperature}{a boolean to answer the question: does your pest or pathogen have a temperature at which it cannot survive? (TRUE or FALSE)}

\item{temperature_file}{path to raster file with temperature data for minimum temperature}

\item{lethal_temperature}{the temperature in degrees C at which lethal temperature related mortality occurs for your pest or pathogen (-50 and 60)}

\item{lethal_temperature_month}{the month in which lethal temperature related mortality occurs for your pest or pathogen integer value between 1 and 12}

\item{mortality_on}{boolean to turn host mortality on and off (TRUE or FALSE)}

\item{mortality_rate}{rate at which mortality occurs value between 0 and 1}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years integer >= 1}

\item{management}{boolean to allow use of managemnet (TRUE or FALSE)}

\item{treatment_dates}{dates in which to apply treatment list with format ('YYYY_MM_DD') (needs to be the same length as treatment_file and pesticide_duration)}

\item{treatments_file}{path to raster file with treatment data by dates (needs to be the same length as treatment_dates and pesticide_duration)}

\item{treatment_method}{what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.}

\item{natural_kernel_type}{what type of dispersal kernel should be used for natural dispersal can be ('cauchy', 'exponential')}

\item{anthropogenic_kernel_type}{what type of dispersal kernel should be used for anthropogenic dispersal can be ('cauchy', 'exponential')}

\item{natural_dir}{sets the predominate direction of natural dispersal usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')}

\item{natural_kappa}{sets the strength of the natural direction in the von-mises distribution numeric value between 0.01 and 12}

\item{anthropogenic_dir}{sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')}

\item{anthropogenic_kappa}{sets the strength of the anthropogenic direction in the von-mises distribution numeric value between 0.01 and 12}

\item{pesticide_duration}{how long does the pestcide (herbicide, vaccine, etc..) last before the host is susceptible again. If value is 0 treatment is a culling (i.e. host removal) not a pesticide treatment. (needs to be the same length as treatment_dates and treatment_file)}

\item{pesticide_efficacy}{how effictive is the pesticide at preventing the disease or killing the pest (if this is 0.70 then when applied it successfully treats 70 percent of the plants or animals)}

\item{mask}{Raster file used to provide a mask to remove 0's that are not true negatives from comparisons (e.g. mask out lakes and oceans from statics if modeling terrestrial species).}

\item{success_metric}{Choose which success metric to use for calibration. Choices are "number of locations", "number of locations and total distance", "number of locations, number of infections, and total distance", or "residual error". Default is "number of locations and total distance"}

\item{output_frequency}{sets when outputs occur either ('year', 'month', 'week', 'day' or 'time step')}

\item{output_frequency_n}{sets number units ('year', 'month', 'week', 'day' or 'time step') in which to export model results.}

\item{movements_file}{this is a csv file with columns lon_from, lat_from, lon_to, lat_to, number of animals, and date.}

\item{use_movements}{this is a boolean to turn on use of the movement module.}

\item{start_exposed}{Do your initial conditions start as exposed or infected (only used if model_type is "SEI")}

\item{generate_stochasticity}{Boolean to indicate whether to use stochasticity in reproductive functions default is TRUE}

\item{establishment_stochasticity}{Boolean to indicate whether to use stochasticity in establishment functions default is TRUE}

\item{movement_stochasticity}{Boolean to indicate whether to use stochasticity in movement functions default is TRUE}

\item{deterministic}{Boolean to indicate whether to use a deterministic dispersal kernel default is FALSE}

\item{establishment_probability}{Threshold to determine establishment if establishment_stochasticity is FALSE (range 0 to 1, default = 0.5)}

\item{dispersal_percentage}{Percentage of dispersal used to calculate the bounding box for deterministic dispersal}

\item{quarantine_areas_file}{path to raster file with quarantine boundaries used in calculating likelihood of quarantine escape if use_quarantine is TRUE}

\item{use_quarantine}{boolean to indicate whether or not there is a quarantine area if TRUE must pass in a file with polygons (geopackage or shapefile) indicating the quarantine areas (default = FALSE)}
}
\value{
a dataframe of the variables saved and their success metrics for each run
}
\description{
Markov Chain Monte Carlo approximation is used to estimate the reproductive rate and the short distance scale parameters. Model 
accuracy is gauged using a custom quantity allocation disagreement function to assess accuracy of spatial configuration. The 
calibration uses this metric to determine if an MCMC run is kept either because it improves the results or randomly gets kept 
despite being worse. We recommend running calibration for at least 10,000 iterations but even more will provide a better result. 
If the model converges and doesn't improve for awhile it will exist calibration prior to reaching the total number of iterations specified.
}
\examples{
\dontrun{
}
}
