% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_qad_lite.R
\name{configuration_qad_lite}
\alias{configuration_qad_lite}
\title{Create PoPS configuration file for quantity allocation and disagreement lite function}
\usage{
configuration_qad_lite(
  reference,
  comparison,
  mask,
  use_configuration = FALSE,
  use_distance = FALSE,
  use_rmse = FALSE,
  output_folder_path = ""
)
}
\arguments{
\item{reference}{Path to the raster with ground truth data. Values > 1 are
reclassified to 1, values < 1 to 0, and NA remains unchanged (except for RMSE).}

\item{comparison}{Path to the raster with simulated data. Values > 1 are
reclassified to 1, and values < 1 to 0 (except for RMSE).}

\item{mask}{Path to the raster mask.}

\item{use_configuration}{Boolean to use configuration disagreement for model
comparison. Default is FALSE.}

\item{use_distance}{Boolean to compare distances between simulations and
observations. Default is FALSE.}

\item{use_rmse}{Boolean to use RMSE (root mean square error) for comparison. Used only if
population data is accurate enough.}

\item{output_folder_path}{this is the full path with either / or \
(e.g., "C:/user_name/desktop/pops_sod_2020_2023/outputs/"). If not provided,
the config_qad_rds, masked and reclassified rasters will be exported to the
same directories as the input reference and comparison folders.}
}
\description{
Pre-process input data for the quantity, allocation, and disagreement lite function.
This function separates pre-processing steps from validation to optimize memory usage.
}
