% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_lite.R
\name{pops_lite}
\alias{pops_lite}
\title{PoPS Lite Model}
\usage{
pops_lite(
  config_file = "",
  number_of_cores = NULL,
  number_of_iterations = NULL,
  random_seed = NULL,
  new_dirs_path = NULL
)
}
\arguments{
\item{config_file}{Path to config file produced when calling `configuration`.
The config file includes all data necessary used to set up c++ PoPS model}

\item{number_of_cores}{(Default = `NULL`) Specify the number of cores to use
for the computation. If a value is provided, it overrides `config$core_count`.
If set to `NULL`, the value from `config$core_count` will be used.}

\item{number_of_iterations}{(Default = `NULL`) Specify the number of iterations
to run the PoPS model. If a value is provided, it overrides
`config$number_of_iterations` in `config_file`.}

\item{random_seed}{Sets the random seed for the simulation used for reproducibility.
Value provided will replace the value in `config_file$random_seed`. If no value provided,
a random value will replace the value in  `config_file$random_seed` for
each PoPS (lite) run iteration.}

\item{new_dirs_path}{(Default = `NULL`) Specify a new root directory to update
input and output file paths in the `config_file`. This helps adapt the file paths
when running the configuration on a different workstation with a different root path,
without needing to recreate the entire `config_file`.
If `new_dirs_path` is `NULL` (default), the original paths in `config_file` are used.
When specified, `new_dirs_path` should point to the top-level folder containing
the input files and output folder. The folder structure under this top-level
directory must match the structure in the original `config_file` If no match is
found, the original input file paths and output folder remain unchanged.}
}
\value{
list of infected and susceptible per year
}
\description{
A process-based model for forecasting the spread of forest/agricultural pests
or pathogens. This model accounts for weather and environmental effects on
reproduction and survival through multiple stochastic simulations. It
propagates uncertainty in parameters, initial conditions, and drivers,
similar to pops_multirun, but directly exports raw outputs from pops_model.
}
\details{
When to Use PoPS Lite:
- To export raw simulation data from the `pops_model` function
- Extent or resolution constraints make `pops_multirun` impractical due to
time or computational constraints.

Dynamic Configuration:
`number_of_cores`, `number_of_iterations`, and `random_seeds` in the
`config_file` can be updated dynamically. See parameter descriptions below
on how to dynamically update these parameters.

Multiple Random Seed handling:
- If multiple_random_seeds = TRUE and file_random_seeds = NULL in the
`config_file`, a new set of random seeds will be sampled and exported as
[unique_id]_forecast_random_seeds.csv in the output folder.
- The same [unique_id] is used for raw output files to link seeds with
their respective runs.
}
