% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_model.R
\name{pops_model}
\alias{pops_model}
\title{PoPS (Pest or Pathogen Spread) model}
\usage{
pops_model(
  random_seed,
  use_multiple_random_seeds,
  random_seeds,
  use_lethal_temperature,
  lethal_temperature,
  lethal_temperature_month,
  use_overwinter_survival,
  overwinter_survival_rate_month,
  overwinter_survival_rate_day,
  host_pools,
  total_populations,
  competency_table,
  pest_host_table,
  mortality_on,
  quarantine_areas,
  quarantine_directions,
  treatment_maps,
  treatment_dates,
  pesticide_durations,
  use_movements,
  movements,
  movements_dates,
  weather,
  temperature,
  survival_rates,
  weather_coefficient,
  weather_coefficient_sd,
  res,
  rows_cols,
  time_step,
  reproductive_rate,
  spatial_indices,
  season_month_start_end,
  soil_reservoirs,
  start_date = "2018-01-01",
  end_date = "2018-12-31",
  treatment_method = "ratio",
  natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy",
  use_anthropogenic_kernel = FALSE,
  percent_natural_dispersal = 0,
  natural_distance_scale = 21,
  anthropogenic_distance_scale = 0,
  natural_dir = "NONE",
  natural_kappa = 0,
  anthropogenic_dir = "NONE",
  anthropogenic_kappa = 0,
  output_frequency = "year",
  output_frequency_n = 1,
  quarantine_frequency = "year",
  quarantine_frequency_n = 1,
  use_quarantine = FALSE,
  spreadrate_frequency = "year",
  spreadrate_frequency_n = 1,
  mortality_frequency = "year",
  mortality_frequency_n = 1,
  use_spreadrates = FALSE,
  model_type_ = "SI",
  latency_period = 0,
  generate_stochasticity = TRUE,
  establishment_stochasticity = TRUE,
  movement_stochasticity = TRUE,
  dispersal_stochasticity = TRUE,
  establishment_probability = 0,
  dispersal_percentage = 0.99,
  use_overpopulation_movements = FALSE,
  overpopulation_percentage = 0,
  leaving_percentage = 0,
  leaving_scale_coefficient = 1,
  bbox = NULL,
  network_min_distances = c(0),
  network_max_distances = c(0),
  network_files = c(""),
  network_movement_types = c("walk"),
  network_weights = c(1),
  weather_size = 0,
  weather_type = "deterministic",
  dispersers_to_soils_percentage = 0,
  use_soils = FALSE
)
}
\arguments{
\item{random_seeds}{vector of random seeds in the order of "disperser_generation",
"natural_dispersal", "anthropogenic_dispersal", "establishment", "weather",
"lethal_temperature", "movement", "overpopulation", "survival_rate", "soil"}

\item{host_pools}{list of host_pool lists with each host_pool list containing matrices of
infected, susceptible, exposed, total_hosts, total_exposed, resistant, mortality, and
mortality_tracker for that host.}

\item{total_populations}{matrix of total populations}

\item{mortality_on}{Boolean to indicate if mortality is used}

\item{quarantine_areas}{areas that are set as quarantined for computing
escape from quarantine statistics.}

\item{treatment_maps}{list of matrices where treatment or management has
occurred in a given year}

\item{use_movements}{this is a boolean to turn on use of the movement module.}

\item{movements}{a matrix with columns lon_from, lat_from, lon_to, lat_to,
number of animals, and date.}

\item{movements_dates}{this is a list of dates passed as strings in the
format 'YYYY-MM-DD'}

\item{weather}{Boolean that is true if weather is used}

\item{temperature}{vector of matrices of temperature values used to check
against lethal temperature}

\item{survival_rates}{vector of matrices of survival rates used to determine percentage of
overwinter population that emerges}

\item{weather_coefficient}{vector of matrices of weather coefficients}

\item{weather_coefficient_sd}{vector of matrices of weather coefficient standard deviations.}

\item{res}{vector of east/west resolution and north/south resolution}

\item{rows_cols}{vector of number of rows and columns in the raster files}

\item{reproductive_rate}{number of spores or pest units produced by a single
host under optimal weather conditions}

\item{spatial_indices}{list of all spatial locations with suitable hosts}

\item{season_month_start_end}{vector of months when spread starts and stops}

\item{soil_reservoirs}{list of matrices with soil pests created from soil_pest_file.}

\item{use_anthropogenic_kernel}{A boolean that turns on the use of the
anthropogenic or long distance dispersal portion of the 2 scale dispersal
kernel function}

\item{percent_natural_dispersal}{what percentage of dispersal is natural
range versus anthropogenic range value between 0 and 1}

\item{natural_distance_scale}{distance scale parameter for natural range
dispersal kernel numeric value > 0}

\item{anthropogenic_distance_scale}{distance scale parameter for
anthropogenic range dispersal kernel numeric value > 0}

\item{natural_kappa}{sets the strength of the natural direction in the
von-mises distribution numeric value between 0.01 and 12}

\item{anthropogenic_kappa}{sets the strength of the anthropogenic direction
in the von-mises distribution numeric value between 0.01 and 12}

\item{quarantine_frequency}{sets how often the quarantine statistics are
calculated either ('year', 'month', 'week', 'day' or 'time step')
(default is year)}

\item{quarantine_frequency_n}{sets number units ('year', 'month', 'week',
'day' or 'time step') in which to calculate and export quarantine statistics.}

\item{spreadrate_frequency}{sets how often the spread rate statistics are
calculated either ('year', 'month', 'week', 'day' or 'time step')
(default is year)}

\item{spreadrate_frequency_n}{sets number units ('year', 'month', 'week',
'day' or 'time step') in which to calculate and export spread rate
statistics.}

\item{model_type_}{What type of model most represents your system. Options
are "SEI" (Susceptible - Exposed - Infected/Infested) or "SI"
(Susceptible - Infected/Infested). Default value is "SI".}

\item{bbox}{bounding box for network kernel}

\item{weather_size}{the number of matrices in a list or layers in a raster object}

\item{weather_type}{string indicating how the weather data is passed in  either
as a mean and standard deviation to represent uncertainty ("probabilistic") or as a time
series ("deterministic")}

\item{dispersers_to_soils_percentage}{range from 0 to 1 representing the percentage
of dispersers that fall to the soil and survive.}
}
\value{
list of vector matrices of infected and susceptible hosts per
simulated year and associated statistics (e.g. spread rate)
}
\description{
Wrapper for pops_model_cpp. A dynamic species distribution model
for pest or pathogen spread in forest or agricultural ecosystems. The model
is process based meaning that it uses understanding of the effect of weather
and other environmental factors on reproduction and survival of the
pest/pathogen in order to forecast spread of the pest/pathogen into the
future.
}
