% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_multirun.R
\name{pops_multirun}
\alias{pops_multirun}
\title{PoPS (Pest or Pathogen Spread) model Multiple Runs}
\usage{
pops_multirun(config)
}
\arguments{
\item{config}{The config with the list of parameters returned from running configuration.}
}
\value{
list of infected and susceptible per year
}
\description{
A dynamic species distribution model for pest or pathogen spread in forest
or agricultural ecosystems. The model is process based meaning that it uses
understanding of the effect of weather and other environmental factors on
reproduction and survival of the pest/pathogen in order to forecast spread
of the pest/pathogen into the future. Run multiple stochastic simulations,
propagating uncertainty in parameters, initial conditions, and drivers.
The model is process based meaning that it uses understanding of the effect
of weather on reproduction and survival of the pest/pathogen in order to
forecast spread of the pest/pathogen into the future.
}
