% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_simulate.R
\name{pops_simulate}
\alias{pops_simulate}
\title{PoPS Simulate}
\usage{
pops_simulate(config_rds_file = "")
}
\arguments{
\item{config_rds_file}{Path to config file produced when calling `configuration`.
The config file includes all data necessary used to set up c++ PoPS model}
}
\value{
list of infected and susceptible per year
}
\description{
A process-based model for forecasting the spread of forest/agricultural pests
or pathogens. This model accounts for weather and environmental effects on
reproduction and survival through multiple stochastic simulations. It
propagates uncertainty in parameters, initial conditions, and drivers,
similar to pops_multirun, but directly exports raw outputs from pops_model.
}
\details{
When to Use PoPS Lite:
- To export raw simulation data from the `pops_model` function
- Extent or resolution constraints make `pops_multirun` impractical due to
time or computational constraints.

Multiple Random Seed handling:
- If use_multiple_random_seeds = TRUE and multiple_random_seeds_file = NULL in the
`config_rds_file`, a new set of random seeds will be sampled and exported as
[unique_id]_forecast_random_seeds.csv in the output folder.
- The same [unique_id] is used for raw output files to link seeds with
their respective runs.
}
