% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_allocation_disagreement.R
\name{quantity_allocation_disagreement}
\alias{quantity_allocation_disagreement}
\title{Quantity Allocation Disagreement}
\usage{
quantity_allocation_disagreement(reference, comparison)
}
\arguments{
\item{reference}{the binary (0 or 1) raster with ground truth data}

\item{comparison}{the binary (0 or 1) raster with simulated data to be compared to ground truth}
}
\value{
data frame of
}
\description{
Quantity Allocation Disagreement
}
