% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_allocation_disagreement_lite.R
\name{quantity_allocation_disagreement_lite}
\alias{quantity_allocation_disagreement_lite}
\title{Calculates quantity and allocation disagreement between two raster datasets.
`quantity_allocation_disagreement_lite` functions similarly to
`quantity_allocation_disagreement` but uses pre-processed inputs from
`config_qad_lite`.}
\usage{
quantity_allocation_disagreement_lite(
  config_qad_lite_file,
  new_dirs_path = NULL,
  use_configuration = NULL,
  use_distance = NULL
)
}
\arguments{
\item{config_qad_lite_file}{the raster with ground truth data. For all metrics expect
RMSE these values are reclassified with values > 1 becoming 1, values < 1}

\item{new_dirs_path}{(Default = `NULL`) Specify a new root directory to update
input and output file paths in the `config_qad_lite_file`. This helps adapt the file paths
when running the configuration on a different workstation with a different root path,
without needing to recreate the entire `config_qad_lite_file`.
If `new_dirs_path` is `NULL` (default), the original paths in `config_qad_lite_file` are used.
When specified, `new_dirs_path` should point to the top-level folder containing
the input files and output folder. The folder structure under this top-level
directory must match the structure in the original `config_qad_lite_file` If no match is
found, the original input file paths and output folder remain unchanged.}

\item{use_configuration}{Boolean if you want to use configuration
disagreement for comparing model runs. NOTE: overrides boolean provided in
the quantity, allocation, and disagreement config file.}

\item{use_distance}{Boolean if you want to compare distance between
simulations and observations. NOTE: overrides boolean provided in
the quantity, allocation, and disagreement config file.}
}
\value{
A data frame with spatial configuration metrics. Particularly
quantity, allocation, and total disagreement,  omission and commission, and
directional disagreement where directional disagreement.
}
\description{
`config_qad_lite` handles pre-processing, preparing the reference and
comparison raster files, and saves them as an RDS file. This RDS is then
used by `quantity_allocation_disagreement_lite` to compute metrics efficiently.
}
\details{
Separating pre-processing from metric calculation minimizes memory usage
in `quantity_allocation_disagreement_lite`, making it suitable for
large extent, fine-resolution rasters.

Uses quantity and allocation disagreement metrics by Pontius and Millones
(2014) and omission and commission errors of comparing a modeled raster
data set to a reference raster data set.
}
