% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_config_paths}
\alias{update_config_paths}
\title{Function to dynamically update the working directory for all file paths in
the configuration rds output}
\usage{
update_config_paths(config, root)
}
\description{
Function to dynamically update the working directory for all file paths in
the configuration rds output
}
