% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validates the accuracy of the calibrated reproductive rate and dispersal
scales of the pops model.}
\usage{
validate(config)
}
\value{
a data frame of statistical measures of model performance.
}
\description{
This function uses the quantity, allocation, and configuration disagreement
to validate the model across the landscape using the parameters from the
calibrate function. Ideally the model is calibrated with 2 or more years of
data and validated for the last year or if you have 6 or more years of data
then the model can be validated for the final 2 years.
}
