% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrate function for pops and pops_model functions}
\usage{
calibrate(infected_years_file, num_interations, start_reproductive_rate,
  start_short_distance_scale, sd_reproductive_rate,
  sd_short_distance_scale, infected_file, host_file, total_plants_file,
  reproductive_rate = 3, use_lethal_temperature = FALSE,
  temp = FALSE, precip = FALSE, management = FALSE,
  mortality_on = FALSE, temperature_file = "",
  temperature_coefficient_file = "",
  precipitation_coefficient_file = "", treatments_file = "",
  season_month_start = 1, season_month_end = 12, time_step = "month",
  start_time = 2018, end_time = 2020, treatment_years = c(0),
  dispersal_kern = "cauchy", percent_short_distance_dispersal = 1,
  short_distance_scale = 59, long_distance_scale = 0,
  lethal_temperature = -12.87, lethal_temperature_month = 1,
  mortality_rate = 0, mortality_time_lag = 0, wind_dir = "NONE",
  kappa = 0)
}
\arguments{
\item{infected_years_file}{}

\item{num_interations}{}

\item{start_reproductive_rate}{}

\item{start_short_distance_scale}{}

\item{sd_reproductive_rate}{}

\item{sd_short_distance_scale}{}

\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster file with number of hosts}

\item{total_plants_file}{path to raster file with number of total plants}

\item{reproductive_rate}{number of spores or pest units produced by a single host under optimal weather conditions}

\item{use_lethal_temperature}{does your pest or pathogen have a temperature at which it cannot survive}

\item{temp}{allows the use of temperature coefficients to modify spread}

\item{precip}{allows the use of precipitation coefficients to modify spread}

\item{management}{boolean to allow use of managemnet}

\item{mortality_on}{boolean to turn host mortality on and off}

\item{temperature_file}{path to raster file with temperature data for minimum temperature}

\item{temperature_coefficient_file}{path to raster file with}

\item{precipitation_coefficient_file}{path to raster file with}

\item{treatments_file}{path to raster file with treatment data by years}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen}

\item{season_month_end}{when does spread end during the year for your pest or pathogen}

\item{time_step}{how often should spread occur}

\item{start_time}{first year to start the simulation}

\item{end_time}{last year of the simulation}

\item{treatment_years}{years in which to apply treatment}

\item{dispersal_kern}{what type of dispersal kernel should be used}

\item{percent_short_distance_dispersal}{what percentage of dispersal is short range versus long range}

\item{short_distance_scale}{distance scale parameter for short range dispersal kernel}

\item{long_distance_scale}{distance scale parameter for long range dispersal kernel}

\item{lethal_temperature}{the temperature at which mortality occurs for your pest or pathogen}

\item{lethal_temperature_month}{the month in which mortality occurs}

\item{mortality_rate}{rate at which mortality occurs}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years}

\item{wind_dir}{sets the wind direction}

\item{kappa}{sets the strength of the wind direction in the von-mises distribution}
}
\value{
a dataframe of the variables saved and their success metrics for each run
}
\description{
Calibrate function for pops and pops_model functions
}
