% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrates the reproductive rate and dispersal scales of the pops model.}
\usage{
calibrate(
  infected_years_file,
  number_of_observations = 1,
  prior_number_of_observations = 0,
  prior_means = c(0, 0, 0, 0, 0, 0),
  prior_cov_matrix = matrix(0, 6, 6),
  params_to_estimate = c(T, T, T, T, F, F),
  number_of_generations = 7,
  generation_size = 1000,
  infected_file,
  host_file,
  total_populations_file,
  temp = FALSE,
  temperature_coefficient_file = "",
  precip = FALSE,
  precipitation_coefficient_file = "",
  model_type = "SI",
  latency_period = 0,
  time_step = "month",
  season_month_start = 1,
  season_month_end = 12,
  start_date = "2008-01-01",
  end_date = "2008-12-31",
  use_lethal_temperature = FALSE,
  temperature_file = "",
  lethal_temperature = -12.87,
  lethal_temperature_month = 1,
  mortality_on = FALSE,
  mortality_rate = 0,
  mortality_time_lag = 0,
  mortality_frequency = "year",
  mortality_frequency_n = 1,
  management = FALSE,
  treatment_dates = c(""),
  treatments_file = "",
  treatment_method = "ratio",
  natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy",
  natural_dir = "NONE",
  natural_kappa = 0,
  anthropogenic_dir = "NONE",
  anthropogenic_kappa = 0,
  pesticide_duration = c(0),
  pesticide_efficacy = 1,
  mask = NULL,
  output_frequency = "year",
  output_frequency_n = 1,
  movements_file = "",
  use_movements = FALSE,
  start_exposed = FALSE,
  generate_stochasticity = TRUE,
  establishment_stochasticity = TRUE,
  movement_stochasticity = TRUE,
  deterministic = FALSE,
  establishment_probability = 0.5,
  dispersal_percentage = 0.99,
  quarantine_areas_file = "",
  use_quarantine = FALSE,
  use_spreadrates = FALSE,
  use_overpopulation_movements = FALSE,
  overpopulation_percentage = 0,
  leaving_percentage = 0,
  leaving_scale_coefficient = 1,
  calibration_method = "ABC",
  number_of_iterations = 1e+05,
  exposed_file = "",
  verbose = TRUE,
  write_outputs = "None",
  output_folder_path = "",
  use_distance = FALSE,
  use_rmse = FALSE
)
}
\arguments{
\item{infected_years_file}{Raster file with years of initial
infection/infestation as individual locations of a pest or pathogen. This is
a multiband raster file (e.g. .tif) with each band representing a unique time
step (e.g. band 1 = year 1 .... band 6 = year 6 or band 1 = week 1 .... band
6 = week 6). This needs to allign with both the time step selection and start
and end dates selection. Units for infections are based on data availability
and the way the units used for your host file creation (e.g. percent area, #
of hosts per cell, etc.). This doesn't include the start year which passed in
in the initial_infected_file (e.g. if we had observation data from 2017,
2018, and 2019 the 2017 raster file would be the initial_infected_file and a
dual band raster file would have band 1 = 2018 and band 2 = 2019 observations)}

\item{number_of_observations}{the number of observations used for this
calibration. Useful if using previous calibration. This is used to
weight the parameters when updating parameters when new data becomes
available. Example if we have 2,000 observations in 2019 and had 1,000
observations in 2018 and 1,000 in 2017, we would use 2,000 here and 2,000 for
our prior_number_of_observations.}

\item{prior_number_of_observations}{the number of total observations from
previous calibrations used to weight the posterior distributions (if this is
a new calibration this value takes the form of a prior weight (0 - 1)). This
is used to weight the parameters when updating parameters when new data
becomes available. Example if we have 2,000 observations in 2019 and had
1,000 observations in 2018 and 1,000 in 2017, we would use 2,000 here and
2,000 for our number_of_observations.}

\item{prior_means}{A vector of the means of your parameters you are
estimating in order from (reproductive_rate, natural_dispersal_distance,
percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa,
and anthropogenic kappa). This is used when updating a parameter set from a
previous calibration using the iterative framework.}

\item{prior_cov_matrix}{A covariance matrix from the previous years
posterior parameter estimation ordered from (reproductive_rate,
natural_dispersal_distance, percent_natural_dispersal,
anthropogenic_dispersal_distance, natural kappa, and anthropogenic kappa).
This is used when updating a parameter set from a previous calibration using
the iterative framework.}

\item{params_to_estimate}{A list of booleans specificing which parameters to
estimate ordered from (reproductive_rate, natural_dispersal_distance,
percent_natural_dispersal, anthropogenic_dispersal_distance, natural kappa,
and anthropogenic kappa)}

\item{number_of_generations}{the number of generations to use to decrease
the uncertainty in the parameter estimation (too many and it will take a
long time, too few and your parameter sets will be too wide). This is an ABC
implementation naming convention but should be set to greater than 7 for
robust calibrations. There is a tradeoff between computational time and model
accuracy the larger this number gets. Usually 7 to 9 is the ideal range.}

\item{generation_size}{how many accepted parameter sets should occur in each
generation. For example if generation size is 1,000 then the simulation runs
until 1,000 model runs are less than the threshold value.
We recommend running at least 1,000 but the greater this number the more
accurate the model parameters selected will be.}

\item{infected_file}{Raster file with initial infections. Units for
infections are based on data availability and the way the units used for your
host file is created (e.g. percent area, # of hosts per cell, etc.).}

\item{host_file}{path to raster files with number of hosts and standard
deviation on those estimates can be based in 3 formats (a single file with
number of hosts, a single file with 2 layers number of hosts and standard
deviation, or two files 1 with number of hosts and the other with standard
deviation of those estimates). The units for this can be of many formats the
two most common that we use are either percent area (0 to 100) or # of hosts
in the cell. Usually depends on data available and estimation methods.}

\item{total_populations_file}{path to raster file with number of total
populations of all hosts and non-hosts. This depends on how your host data is
set up. If host is percent area then this should be a raster with values
that are 100 anywhere with host. If host file is # of hosts in a cell then
this should be a raster with values that are the max of the host raster any
where the # of hosts is greater than 0.}

\item{temp}{boolean that allows the use of temperature coefficients to
modify spread (TRUE or FALSE)}

\item{temperature_coefficient_file}{path to raster file with temperature
coefficient data for the timestep and and time period specified (e.g. if timestep
= week and start_date = 2017_01_01 and end_date = 2019_12_31 this file would
have 52 * 3 bands = 156 bands with data being weekly precipitation
coefficients). We convert raw precipitation values to coefficients that
affect the reproduction and survival of the pest all values in the raster are
between 0 and 1.}

\item{precip}{boolean that allows the use of precipitation coefficients to
modify spread (TRUE or FALSE)}

\item{precipitation_coefficient_file}{Raster file with precipitation
coefficient data for the timestep and time period specified (e.g. if timestep
= week and start_date = 2017_01_01 and end_date = 2019_12_31 this file would
have 52 * 3 bands = 156 bands with data being weekly precipitation
coefficients). We convert raw precipitation values to coefficients that
affect the reproduction and survival of the pest all values in the raster are
between 0 and 1.}

\item{model_type}{What type of model most represents your system. Options
are "SEI" (Susceptible - Exposed - Infected/Infested) or "SI"
(Susceptible - Infected/Infested). Default value is "SI".}

\item{latency_period}{How many times steps does it take to for exposed
populations become infected/infested. This is an integer value and must be
greater than 0 if model type is SEI.}

\item{time_step}{how often should spread occur options: ('day', 'week',
'month').}

\item{season_month_start}{when does spread first start occurring in the year
for your pest or pathogen (integer value between 1 and 12)}

\item{season_month_end}{when does spread end during the year for your pest
or pathogen (integer value between 1 and 12)}

\item{start_date}{date to start the simulation with format ('YYYY_MM_DD')}

\item{end_date}{date to end the simulation with format ('YYYY_MM_DD')}

\item{use_lethal_temperature}{a boolean to answer the question: does your
pest or pathogen have a temperature at which it cannot survive?
(TRUE or FALSE)}

\item{temperature_file}{path to raster file with temperature data for
minimum temperature}

\item{lethal_temperature}{the temperature in degrees C at which lethal
temperature related mortality occurs for your pest or pathogen (-50 and 60)}

\item{lethal_temperature_month}{the month in which lethal temperature
related mortality occurs for your pest or pathogen integer value between 1
and 12}

\item{mortality_on}{boolean to turn host mortality on and off
(TRUE or FALSE)}

\item{mortality_rate}{rate at which mortality occurs value between 0 and 1}

\item{mortality_time_lag}{time lag from infection until mortality can occur
in time steps integer >= 1}

\item{mortality_frequency}{sets the frequency of mortality calculations occur
either ('year', 'month', week', 'day', 'time step', or 'every_n_steps')}

\item{mortality_frequency_n}{sets number of units from mortality_frequency in
which to run the mortality calculation if mortality_frequency is
'every_n_steps'. Must be an integer >= 1.}

\item{management}{boolean to allow use of management (TRUE or FALSE)}

\item{treatment_dates}{dates in which to apply treatment list with format
('YYYY_MM_DD') (needs to be the same length as treatment_file and
pesticide_duration)}

\item{treatments_file}{path to raster files with treatment data by dates.
Needs to be a list of files the same length as treatment_dates and
pesticide_duration.}

\item{treatment_method}{what method to use when applying treatment one of
("ratio" or "all infected"). ratio removes a portion of all infected and
susceptibles, all infected removes all infected a portion of susceptibles.}

\item{natural_kernel_type}{what type of dispersal kernel should be used for
natural dispersal. Current dispersal kernel options are
 ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law',
 'hyperbolic secant', 'gamma', 'weibull',
 'logistic')}

\item{anthropogenic_kernel_type}{what type of dispersal kernel should be
used for anthropogenic dispersal. Current dispersal kernel options are
 ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law',
 'hyperbolic secant', 'gamma', 'weibull',
 'logistic')}

\item{natural_dir}{sets the predominate direction of natural dispersal
usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE',
'NONE')}

\item{natural_kappa}{sets the strength of the natural direction in the
von-mises distribution numeric value between 0.01 and 12}

\item{anthropogenic_dir}{sets the predominate direction of anthropogenic
dispersal usually due to human movement typically over long distances (e.g.
nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S',
'SE', 'E', 'NE', 'NONE')}

\item{anthropogenic_kappa}{sets the strength of the anthropogenic direction
in the von-mises distribution numeric value between 0.01 and 12}

\item{pesticide_duration}{how long does the pestcide (herbicide, vaccine,
etc..) last before the host is susceptible again. If value is 0 treatment
is a culling (i.e. host removal) not a pesticide treatment. (needs to be the
 same length as treatment_dates and treatment_file)}

\item{pesticide_efficacy}{how effictive is the pesticide at preventing the
disease or killing the pest (if this is 0.70 then when applied it
successfully treats 70 percent of the plants or animals)}

\item{mask}{Raster file used to provide a mask to remove 0's that are not
true negatives from comparisons (e.g. mask out lakes and oceans from statics
if modeling terrestrial species). A numerical value represents the area you
want to calculate statistics on and an NA value represents the area to remove
from the statistics.}

\item{output_frequency}{sets when outputs occur either ('year', 'month',
'week', 'day', 'time step', or 'every_n_steps')}

\item{output_frequency_n}{sets number of units from output_frequency in which
to export model results if mortality_frequency is 'every_n_steps'.
Must be an integer >= 1.}

\item{movements_file}{this is a csv file with columns lon_from, lat_from,
lon_to, lat_to, number of animals, and date.}

\item{use_movements}{this is a boolean to turn on use of the movement module.}

\item{start_exposed}{Do your initial conditions start as exposed or infected
(only used if model_type is "SEI"). Default False. If this is TRUE need to
have both an infected_file (this can be a raster of all 0's) and exposed_file}

\item{generate_stochasticity}{Boolean to indicate whether to use
stochasticity in reproductive functions default is TRUE}

\item{establishment_stochasticity}{Boolean to indicate whether to use
stochasticity in establishment functions default is TRUE}

\item{movement_stochasticity}{Boolean to indicate whether to use
stochasticity in movement functions default is TRUE}

\item{deterministic}{Boolean to indicate whether to use a deterministic
dispersal kernel default is FALSE}

\item{establishment_probability}{Threshold to determine establishment if
establishment_stochasticity is FALSE (range 0 to 1, default = 0.5)}

\item{dispersal_percentage}{Percentage of dispersal used to calculate the
bounding box for deterministic dispersal}

\item{quarantine_areas_file}{path to raster file with quarantine boundaries
used in calculating likelihood of quarantine escape if use_quarantine is
TRUE}

\item{use_quarantine}{boolean to indicate whether or not there is a
quarantine area if TRUE must pass in a raster file indicating the
quarantine areas (default = FALSE)}

\item{use_spreadrates}{boolean to indicate whether or not to calculate
spread rates}

\item{use_overpopulation_movements}{boolean to indicate whether to use
the overpopulation pest movement module (driven by the natural kernel with
its scale parameter modified by a coefficient)}

\item{overpopulation_percentage}{percentage of occupied hosts when the cell
is considered to be overpopulated}

\item{leaving_percentage}{percentage of pests leaving an overpopulated cell}

\item{leaving_scale_coefficient}{coefficient to multiply scale parameter of
the natural kernel (if applicable)}

\item{calibration_method}{choose which method of calibration to use either
'ABC' (Approximate Bayesian Computation) or 'MCMC' (Markov Chain Monte Carlo
Approximation)}

\item{number_of_iterations}{how many iterations do you want to run to allow
the calibration to converge (recommend a minimum of at least 100,000 but
preferably 1 million).}

\item{exposed_file}{a file with the exposed for the current}

\item{verbose}{Boolean with true printing current status of calibration,
(e.g. the current generation, current particle, and the acceptance rate).
Defaults if FALSE.}

\item{write_outputs}{Either c("summary_outputs", or "None"). If not
"None" output folder path must be provided.}

\item{output_folder_path}{this is the full path with either / or \\ (e.g.,
"C:/user_name/desktop/pops_sod_2020_2023/outputs/")}

\item{use_distance}{Boolean if you want to compare distance between
simulations and observations. Default is FALSE.}

\item{use_rmse}{Boolean if you want to calibrate based on rmse. This is
useful if you have very good population level observations. Default is FALSE.}
}
\value{
a dataframe of the variables saved and their success metrics for
each run
}
\description{
Either Approximate Bayesian Computation or Markov Chain Monte Carlo
Approximation is used to estimate relevant model parameters. Model accuracy
is gauged using a custom quantity allocation disagreement function to assess
accuracy of spatial configuration. We test number of predictions, number of
predicted locations, cumulative distance to nearest infection. The
calibration uses these metrics to determine if a run is kept if it is under
a threshold. either because it improves the results or randomly gets kept
despite being worse. We recommend running calibration for at least 10,000
iterations but even more will provide a better result. If the model converges
and doesn't improve for awhile it will exist calibration prior to reaching
the total number of iterations specified.
}
