% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{Calibrates the reproductive rate and dispersal scales of the pops model.}
\usage{
calibrate(infected_years_file, num_iterations, start_reproductive_rate,
  number_of_cores = NA, start_natural_distance_scale,
  sd_reproductive_rate, sd_natural_distance_scale, infected_file,
  host_file, total_plants_file, temp = FALSE,
  temperature_coefficient_file = "", precip = FALSE,
  precipitation_coefficient_file = "", time_step = "month",
  reproductive_rate = 3, season_month_start = 1,
  season_month_end = 12, start_time = 2018, end_time = 2020,
  use_lethal_temperature = FALSE, temperature_file = "",
  lethal_temperature = -12.87, lethal_temperature_month = 1,
  mortality_on = FALSE, mortality_rate = 0, mortality_time_lag = 0,
  management = FALSE, treatment_dates = c(0), treatments_file = "",
  treatment_method = "ratio", percent_natural_dispersal = 1,
  natural_kernel_type = "cauchy", anthropogenic_kernel_type = "cauchy",
  natural_distance_scale = 21, anthropogenic_distance_scale = 0,
  natural_dir = "NONE", natural_kappa = 0,
  anthropogenic_dir = "NONE", anthropogenic_kappa = 0,
  pesticide_duration = c(0), pesticide_efficacy = 1, mask = NULL,
  success_metric = "quantity")
}
\arguments{
\item{infected_years_file}{Raster file with years of initial infection/infestation as individual locations of a pest or pathogen}

\item{num_iterations}{how many iterations do you want to run to allow the calibration to converge (recommend a minimum of at least 100,000 but preferably 1 million).}

\item{start_reproductive_rate}{starting reproductive rate for MCMC calibration (affects how quickly a series converges) numeric value > 0}

\item{number_of_cores}{number of cores to use for calibration (defaults to the number of cores available on the machine) integer value >= 1}

\item{start_natural_distance_scale}{starting short distance scale parameter for MCMC calibration (affects how quickly a series converges) numeric value > 0}

\item{sd_reproductive_rate}{starting standard deviation for reproductive rate for MCMC calibration (can affect how quickly and if a series converges) numeric value > 0}

\item{sd_natural_distance_scale}{starting standard deviation for short distance scale for MCMC calibration (can affect how quickly and if a series converges) numeric value > 0}

\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster file with number of hosts}

\item{total_plants_file}{path to raster file with number of total plants}

\item{temp}{boolean that allows the use of temperature coefficients to modify spread (TRUE or FALSE)}

\item{temperature_coefficient_file}{path to raster file with temperature coefficient data for the timestep and number of years specified}

\item{precip}{boolean that allows the use of precipitation coefficients to modify spread (TRUE or FALSE)}

\item{precipitation_coefficient_file}{path to raster file with precipitation coefficient data for the timestep and number of years specified}

\item{time_step}{how often should spread occur options: ('day', 'week', 'month')}

\item{reproductive_rate}{number of spores or pest units produced by a single host under optimal weather conditions}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen (integer value between 1 and 12)}

\item{season_month_end}{when does spread end during the year for your pest or pathogen (integer value between 1 and 12)}

\item{start_time}{first year to start the simulation (needs to be a 4 digit year)}

\item{end_time}{last year of the simulation (needs to be a 4 digit year)}

\item{use_lethal_temperature}{a boolean to answer the question: does your pest or pathogen have a temperature at which it cannot survive? (TRUE or FALSE)}

\item{temperature_file}{path to raster file with temperature data for minimum temperature}

\item{lethal_temperature}{the temperature in degrees C at which lethal temperature related mortality occurs for your pest or pathogen (-50 and 60)}

\item{lethal_temperature_month}{the month in which lethal temperature related mortality occurs for your pest or pathogen integer value between 1 and 12}

\item{mortality_on}{boolean to turn host mortality on and off (TRUE or FALSE)}

\item{mortality_rate}{rate at which mortality occurs value between 0 and 1}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years integer >= 1}

\item{management}{boolean to allow use of managemnet (TRUE or FALSE)}

\item{treatment_dates}{dates in which to apply treatment list with format (YYYY_MM_DD)}

\item{treatments_file}{path to raster file with treatment data by dates}

\item{treatment_method}{what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.}

\item{percent_natural_dispersal}{what percentage of dispersal is natural range versus anthropogenic range value between 0 and 1}

\item{natural_kernel_type}{what type of dispersal kernel should be used for natural dispersal can be ('cauchy', 'exponential')}

\item{anthropogenic_kernel_type}{what type of dispersal kernel should be used for anthropogenic dispersal can be ('cauchy', 'exponential')}

\item{natural_distance_scale}{distance scale parameter for natural range dispersal kernel numeric value > 0}

\item{anthropogenic_distance_scale}{distance scale parameter for anthropogenic range dispersal kernel numeric value > 0}

\item{natural_dir}{sets the predominate direction of natural dispersal usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')}

\item{natural_kappa}{sets the strength of the natural direction in the von-mises distribution numeric value between 0.01 and 12}

\item{anthropogenic_dir}{sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')}

\item{anthropogenic_kappa}{sets the strength of the anthropogenic direction in the von-mises distribution numeric value between 0.01 and 12}

\item{pesticide_duration}{how long does the pestcide (herbicide, vaccine, etc..) last before the host is susceptible again. If value is 0 treatment is a culling (i.e. host removal) not a pesticide treatment.}

\item{pesticide_efficacy}{how effictive is the pesticide at preventing the disease or killing the pest (if this is 0.70 then when applied it successfully treats 70 percent of the plants or animals)}

\item{mask}{Raster file used to provide a mask to remove 0's that are not true negatives from comparisons (e.g. mask out lakes and oceans from statics if modeling terrestrial species).}

\item{success_metric}{Choose which success metric to use for calibration. Choices are "quantity", "quantity and configuration", "residual error" and "odds ratio". Default is "quantity"}
}
\value{
a dataframe of the variables saved and their success metrics for each run
}
\description{
Markov Chain Monte Carlo approximation is used to estimate the reproductive rate and the short distance scale parameters. Model 
accuracy is gauged using a custom quantity allocation disagreement function to assess accuracy of spatial configuration. The 
calibration uses this metric to determine if an MCMC run is kept either because it improves the results or randomly gets kept 
despite being worse. We recommend running calibration for at least 10,000 iterations but even more will provide a better result. 
If the model converges and doesn't improve for awhile it will exist calibration prior to reaching the total number of iterations specified.
}
\examples{
\dontrun{
}
}
