% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{configuration}
\alias{configuration}
\title{PoPS (configuration}
\usage{
configuration(config_file, testing = FALSE)
}
\arguments{
\item{config_file}{yaml or csv file with paths and data necessary for formatting data to
be used in the c++ model}

\item{testing}{only used during tests otherwise ignored}
}
\value{
config list with all data ready for pops C++ or error message
}
\description{
Function with a single input and output list for parsing, transforming,
and performing all checks for all functions to run the pops c++ model
}
