% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{configuration}
\alias{configuration}
\title{PoPS (configuration}
\usage{
configuration(config)
}
\arguments{
\item{config}{list of all data necessary used to set up c++ model}
}
\value{
list of infected and susceptible per year
}
\description{
Function with a single input and output list for parsing, transforming,
and performing all checks for all functions to run the pops c++ model
}
