% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_summary_stats_and_stacks.R
\name{create_summary_stats_and_stacks}
\alias{create_summary_stats_and_stacks}
\title{Create summary stats and raster summaries}
\usage{
create_summary_stats_and_stacks(config)
}
\arguments{
\item{config}{the config used for running PoPS lite for the simulations.}
}
\value{
creates and writes raster stacks from all pops_lite runs in the outputs folder. Also
reates and writes mean, standard deviation, median, min, and max runs, and summary statistics.
}
\description{
This function takes the outputs from PoPS lite and calculates the mean, max, min, and standard
deviation of each cell and calculates the summary statistics for the study area.
}
