% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops.r
\name{pops}
\alias{pops}
\title{PoPS (Pest or Pathogen Spread) model}
\usage{
pops(infected_file, host_file, total_plants_file, reproductive_rate = 3,
  use_lethal_temperature = FALSE, temp = FALSE, precip = FALSE,
  temperature_file = "", temperature_coefficient_file = "",
  precipitation_coefficient_file = "", season_month_start = 1,
  season_month_end = 12, time_step = "month", start_time = 2018,
  end_time = 2022, dispersal_kern = "cauchy",
  percent_short_distance_dispersal = 1, short_distance_scale = 59,
  long_distance_scale = 0, lethal_temperature = -12.87,
  lethal_temperature_month = 1, wind_dir = "NONE", kappa = 0,
  random_seed = 42)
}
\arguments{
\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster file with number of hosts}

\item{total_plants_file}{path to raster file with number of total plants}

\item{reproductive_rate}{}

\item{use_lethal_temperature}{}

\item{temp}{}

\item{precip}{}

\item{temperature_file}{path to raster file with temperature data for minimum temperature}

\item{temperature_coefficient_file}{path to raster file with}

\item{precipitation_coefficient_file}{path to raster file with}

\item{season_month_start}{}

\item{season_month_end}{}

\item{time_step}{}

\item{start_time}{}

\item{end_time}{}

\item{dispersal_kern}{}

\item{percent_short_distance_dispersal}{}

\item{short_distance_scale}{}

\item{long_distance_scale}{}

\item{lethal_temperature}{}

\item{lethal_temperature_month}{}

\item{wind_dir}{sets the wind direction}

\item{kappa}{sets the strength of the wind direction in the von-mises distribution}

\item{random_seed}{sets the random seed for the simulation used for reproducibility}
}
\description{
PoPS (Pest or Pathogen Spread) model
}
\examples{

}
