% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops.r
\name{pops}
\alias{pops}
\title{PoPS (Pest or Pathogen Spread) model}
\usage{
pops(config)
}
\arguments{
\item{config}{The config with the list of parameters returned from running configuration.}
}
\value{
list of infected and susceptible per year
}
\description{
A dynamic species distribution model for pest or pathogen spread
in forest or agricultural ecosystems. The model is process based meaning that
it uses understanding of the effect of weather and other environmental
factors on reproduction and survival of the pest/pathogen in order to
forecast spread of the pest/pathogen into the future. This function performs
a single stochastic realization of the model and is predominantly used for
automated tests of model features.
}
