% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops.r
\name{pops}
\alias{pops}
\title{PoPS (Pest or Pathogen Spread) model}
\usage{
pops(infected_file, host_file, total_plants_file, reproductive_rate = 3,
  use_lethal_temperature = FALSE, temp = FALSE, precip = FALSE,
  management = FALSE, mortality_on = FALSE, temperature_file = "",
  temperature_coefficient_file = "",
  precipitation_coefficient_file = "", treatments_file = "",
  season_month_start = 1, season_month_end = 12, time_step = "month",
  start_time = 2018, end_time = 2020, treatment_years = c(0),
  dispersal_kern = "cauchy", percent_short_distance_dispersal = 1,
  short_distance_scale = 59, long_distance_scale = 0,
  lethal_temperature = -12.87, lethal_temperature_month = 1,
  mortality_rate = 0, mortality_time_lag = 0,
  treatment_method = "ratio", treatment_date = 12, wind_dir = "NONE",
  kappa = 0, random_seed = NULL)
}
\arguments{
\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster file with number of hosts}

\item{total_plants_file}{path to raster file with number of total plants}

\item{reproductive_rate}{number of spores or pest units produced by a single host under optimal weather conditions}

\item{use_lethal_temperature}{does your pest or pathogen have a temperature at which it cannot survive}

\item{temp}{allows the use of temperature coefficients to modify spread}

\item{precip}{allows the use of precipitation coefficients to modify spread}

\item{management}{boolean to allow use of managemnet}

\item{mortality_on}{boolean to turn host mortality on and off}

\item{temperature_file}{path to raster file with temperature data for minimum temperature}

\item{temperature_coefficient_file}{path to raster file with}

\item{precipitation_coefficient_file}{path to raster file with}

\item{treatments_file}{path to raster file with treatment data by years}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen}

\item{season_month_end}{when does spread end during the year for your pest or pathogen}

\item{time_step}{how often should spread occur}

\item{start_time}{first year to start the simulation}

\item{end_time}{last year of the simulation}

\item{treatment_years}{years in which to apply treatment}

\item{dispersal_kern}{what type of dispersal kernel should be used}

\item{percent_short_distance_dispersal}{what percentage of dispersal is short range versus long range}

\item{short_distance_scale}{distance scale parameter for short range dispersal kernel}

\item{long_distance_scale}{distance scale parameter for long range dispersal kernel}

\item{lethal_temperature}{the temperature at which mortality occurs for your pest or pathogen}

\item{lethal_temperature_month}{the month in which mortality occurs}

\item{mortality_rate}{rate at which mortality occurs}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years}

\item{treatment_method}{what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.}

\item{treatment_date}{the time during the year that treatment is applied. Currently monthly option so can be 1 -12. Default is 12.}

\item{wind_dir}{sets the wind direction}

\item{kappa}{sets the strength of the wind direction in the von-mises distribution}

\item{random_seed}{sets the random seed for the simulation used for reproducibility}
}
\value{
list of infected and susceptible per year
}
\description{
A dynamic species distribution model for pest or pathogen spread in forest or agricultural ecosystems. The model is process based
meaning that it uses understanding of the effect of weather on reproduction and survival of the pest/pathogen in order to forecast
spread of the pest/pathogen into the future.
}
\examples{
infected_file <-  system.file("extdata", "SODexample", "initial_infection2001.tif", 
package = "PoPS")
host_file <- system.file("extdata", "SODexample", "host.tif", package = "PoPS")
total_plants_file <- system.file("extdata", "SODexample", "all_plants.tif", package = "PoPS")
temperature_coefficient_file <- system.file("extdata", "SODexample", "weather.tif", package = "
PoPS")
treatments_file <- system.file("extdata", "SODexample", "management.tif", package = "PoPS")

data <- pops(infected_file, host_file, total_plants_file, reproductive_rate = 1.0,
use_lethal_temperature = FALSE, temp = TRUE, precip = FALSE, management = TRUE, 
mortality_on = TRUE, temperature_file = "", temperature_coefficient_file, 
precipitation_coefficient_file ="", treatments_file,
season_month_start = 1, season_month_end = 12, time_step = "week",
start_time = 2001, end_time = 2005, treatment_years = c(2001,2002,2003,2004,2005),
dispersal_kern = "cauchy", percent_short_distance_dispersal = 1.0,
short_distance_scale = 20.57, long_distance_scale = 0.0,
lethal_temperature = -12.87, lethal_temperature_month = 1,
mortality_rate = 0.05, mortality_time_lag = 2,
treatment_date = 12, wind_dir = "NONE", kappa = 0, random_seed = NULL)

}
