% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops.r
\name{pops}
\alias{pops}
\title{PoPS (Pest or Pathogen Spread) model}
\usage{
pops(infected_file, host_file, total_plants_file, temp = FALSE,
  temperature_coefficient_file = "", precip = FALSE,
  precipitation_coefficient_file = "", time_step = "month",
  reproductive_rate = 3, season_month_start = 1,
  season_month_end = 12, start_time = 2018, end_time = 2020,
  use_lethal_temperature = FALSE, temperature_file = "",
  lethal_temperature = -12.87, lethal_temperature_month = 1,
  mortality_on = FALSE, mortality_rate = 0, mortality_time_lag = 0,
  management = FALSE, treatment_dates = c("2000-12-24"),
  treatments_file = "", treatment_method = "ratio",
  percent_natural_dispersal = 1, natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy", natural_distance_scale = 21,
  anthropogenic_distance_scale = 0, natural_dir = "NONE",
  natural_kappa = 0, anthropogenic_dir = "NONE",
  anthropogenic_kappa = 0, pesticide_duration = c(0),
  pesticide_efficacy = 1, random_seed = NULL)
}
\arguments{
\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster file with number of hosts}

\item{total_plants_file}{path to raster file with number of total plants}

\item{temp}{boolean that allows the use of temperature coefficients to modify spread (TRUE or FALSE)}

\item{temperature_coefficient_file}{path to raster file with temperature coefficient data for the timestep and number of years specified}

\item{precip}{boolean that allows the use of precipitation coefficients to modify spread (TRUE or FALSE)}

\item{precipitation_coefficient_file}{path to raster file with precipitation coefficient data for the timestep and number of years specified}

\item{time_step}{how often should spread occur options: ('day', 'week', 'month')}

\item{reproductive_rate}{number of spores or pest units produced by a single host under optimal weather conditions}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen (integer value between 1 and 12)}

\item{season_month_end}{when does spread end during the year for your pest or pathogen (integer value between 1 and 12)}

\item{start_time}{first year to start the simulation (needs to be a 4 digit year)}

\item{end_time}{last year of the simulation (needs to be a 4 digit year)}

\item{use_lethal_temperature}{a boolean to answer the question: does your pest or pathogen have a temperature at which it cannot survive? (TRUE or FALSE)}

\item{temperature_file}{path to raster file with temperature data for minimum temperature}

\item{lethal_temperature}{the temperature in degrees C at which lethal temperature related mortality occurs for your pest or pathogen (-50 and 60)}

\item{lethal_temperature_month}{the month in which lethal temperature related mortality occurs for your pest or pathogen integer value between 1 and 12}

\item{mortality_on}{boolean to turn host mortality on and off (TRUE or FALSE)}

\item{mortality_rate}{rate at which mortality occurs value between 0 and 1}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years integer >= 1}

\item{management}{boolean to allow use of managemnet (TRUE or FALSE)}

\item{treatment_dates}{dates in which to apply treatment list with format (YYYY_MM_DD)}

\item{treatments_file}{path to raster file with treatment data by dates}

\item{treatment_method}{what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.}

\item{percent_natural_dispersal}{what percentage of dispersal is natural range versus anthropogenic range value between 0 and 1}

\item{natural_kernel_type}{what type of dispersal kernel should be used for natural dispersal can be ('cauchy', 'exponential')}

\item{anthropogenic_kernel_type}{what type of dispersal kernel should be used for anthropogenic dispersal can be ('cauchy', 'exponential')}

\item{natural_distance_scale}{distance scale parameter for natural range dispersal kernel numeric value > 0}

\item{anthropogenic_distance_scale}{distance scale parameter for anthropogenic range dispersal kernel numeric value > 0}

\item{natural_dir}{sets the predominate direction of natural dispersal usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')}

\item{natural_kappa}{sets the strength of the natural direction in the von-mises distribution numeric value between 0.01 and 12}

\item{anthropogenic_dir}{sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE', 'NONE')}

\item{anthropogenic_kappa}{sets the strength of the anthropogenic direction in the von-mises distribution numeric value between 0.01 and 12}

\item{pesticide_duration}{how long does the pestcide (herbicide, vaccine, etc..) last before the host is susceptible again. If value is 0 treatment is a culling (i.e. host removal) not a pesticide treatment.}

\item{pesticide_efficacy}{how effictive is the pesticide at preventing the disease or killing the pest (if this is 0.70 then when applied it successfully treats 70 percent of the plants or animals)}

\item{random_seed}{sets the random seed for the simulation used for reproducibility}
}
\value{
list of infected and susceptible per year
}
\description{
A dynamic species distribution model for pest or pathogen spread in forest or agricultural ecosystems. The model is process based
meaning that it uses understanding of the effect of weather on reproduction and survival of the pest/pathogen in order to forecast
spread of the pest/pathogen into the future.
}
\examples{
\dontrun{
infected_file <-  system.file("extdata", "SODexample", "initial_infection2001.tif", 
package = "PoPS")
host_file <- system.file("extdata", "SODexample", "host.tif", package = "PoPS")
total_plants_file <- system.file("extdata", "SODexample", "all_plants.tif", package = "PoPS")
temperature_coefficient_file <- system.file("extdata", "SODexample", "weather.tif", package = "
PoPS")
treatments_file <- system.file("extdata", "SODexample", "management.tif", package = "PoPS")

data <- pops(infected_file, host_file, total_plants_file, reproductive_rate = 1.0,
use_lethal_temperature = FALSE, temp = TRUE, precip = FALSE, management = TRUE, 
mortality_on = TRUE, temperature_file = "", temperature_coefficient_file, 
precipitation_coefficient_file ="", treatments_file,
season_month_start = 1, season_month_end = 12, time_step = "week",
start_time = 2001, end_time = 2005, treatment_dates = c('2001-12-24'),
natural_kernel_type = "cauchy", percent_natural_dispersal = 1.0,
natural_distance_scale = 20.57, anthropogenic_distance_scale = 0.0,
lethal_temperature = -12.87, lethal_temperature_month = 1,
mortality_rate = 0.05, mortality_time_lag = 2,
treatment_date = 12, natural_dir = "NONE", kappa = 0, random_seed = NULL)
}

}
