% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pops_model.R
\name{pops_model}
\alias{pops_model}
\title{PoPS (Pest or Pathogen Spread) model}
\usage{
pops_model(config)
}
\arguments{
\item{config}{a list of all of the necessary data, parameters, files, and paths formatted for
the C++ model created from configuration and modified slightly by the other functions for each
simulation.}
}
\value{
list of vector matrices of infected and susceptible hosts per
simulated year and associated statistics (e.g. spread rate)
}
\description{
Wrapper for pops_model_cpp only used internally in other functions.
}
