% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pops_model}
\alias{pops_model}
\title{PoPS c++ model}
\usage{
pops_model(random_seed, use_lethal_temperature, lethal_temperature,
  lethal_temperature_month, infected, susceptible, total_plants,
  mortality_on, mortality_tracker, mortality, treatment_maps,
  treatment_years, weather, temperature, weather_coefficient, ew_res,
  ns_res, num_rows, num_cols, time_step, reproductive_rate,
  mortality_rate = 0, mortality_time_lag = 2L,
  season_month_start = 1L, season_month_end = 12L, start_time = 2018,
  end_time = 2018, treatment_month = 12L, treatment_method = "ratio",
  natural_kernel_type = "cauchy", anthropogenic_kernel_type = "cauchy",
  use_anthropogenic_kernel = FALSE, percent_natural_dispersal = 0,
  natural_distance_scale = 21, anthropogenic_distance_scale = 0,
  natural_dir = "NONE", natural_kappa = 0,
  anthropogenic_dir = "NONE", anthropogenic_kappa = 0)
}
\arguments{
\item{random_seed}{sets the random seed for the simulation used for reproducibility}

\item{use_lethal_temperature}{does your pest or pathogen have a temperature at which it cannot survive}

\item{lethal_temperature}{the temperature at which lethal temperature related mortality occurs for your pest or pathogen}

\item{lethal_temperature_month}{the month in which lethal temperature related mortality occurs for your pest or pathogen}

\item{infected}{matrix of infected hosts}

\item{susceptible}{matrix of susceptible hosts}

\item{total_plants}{matrix of total plants}

\item{mortality_on}{boolean to turn host mortality on and off}

\item{mortality_tracker}{matrix of 0's to track mortality per year}

\item{mortality}{matrix to track cumulative mortality}

\item{treatment_maps}{list of matrices where treatment or management has occurred in a given year}

\item{treatment_years}{years in which to apply treatment}

\item{weather}{Boolean that is true if weather is used}

\item{temperature}{vector of matrices of temperature values used to check against lethal temperature}

\item{weather_coefficient}{vector of matrices of weather coefficients}

\item{ew_res}{east and west resolution}

\item{ns_res}{north and south resolution}

\item{time_step}{how often should spread occur}

\item{reproductive_rate}{number of spores or pest units produced by a single host under optimal weather conditions}

\item{mortality_rate}{rate at which mortality occurs}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen}

\item{season_month_end}{when does spread end during the year for your pest or pathogen}

\item{start_time}{first year to start the simulation}

\item{end_time}{last year of the simulation}

\item{treatment_month}{the time during the year that treatment is applied. Currently monthly option so can be 1 -12. Default is 12.}

\item{treatment_method}{what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.}

\item{natural_kernel_type}{what type of dispersal kernel should be used for natural dispersal}

\item{anthropogenic_kernel_type}{what type of dispersal kernel should be used for anthropogenic dispersal}

\item{percent_natural_dispersal}{what percentage of dispersal is natural range versus anthropogenic range}

\item{natural_distance_scale}{distance scale parameter for natural range dispersal kernel}

\item{anthropogenic_distance_scale}{distance scale parameter for anthropogenic range dispersal kernel}

\item{natural_dir}{sets the predominate direction of natural dispersal usually due to wind}

\item{natural_kappa}{sets the strength of the natural direction in the von-mises distribution}

\item{anthropogenic_dir}{sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..)}

\item{anthropogenic_kappa}{sets the strength of the anthropogenic direction in the von-mises distribution}
}
\value{
list of vector matrices of infected and suceptible hosts per simulated year
}
\description{
PoPS c++ model
}
