% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pops_model}
\alias{pops_model}
\title{PoPS c++ model}
\usage{
pops_model(random_seed, lethal_temperature, use_lethal_temperature,
  lethal_temperature_month, reproductive_rate, weather, mortality_on,
  short_distance_scale, infected, susceptible, mortality_tracker,
  mortality, total_plants, treatment_maps, treatment_years, temperature,
  weather_coefficient, ew_res, ns_res, time_step, mortality_rate = 0,
  mortality_time_lag = 2L, season_month_start = 1L,
  season_month_end = 12L, start_time = 2018, end_time = 2018,
  dispersal_kern = "cauchy", percent_short_distance_dispersal = 0,
  long_distance_scale = 0, treatment_method = "ratio",
  treatment_month = 12L, wind_dir = "NONE", kappa = 0)
}
\arguments{
\item{random_seed}{sets the random seed for the simulation used for reproducibility}

\item{lethal_temperature}{the temperature at which mortality occurs for your pest or pathogen}

\item{use_lethal_temperature}{does your pest or pathogen have a temperature at which it cannot survive}

\item{lethal_temperature_month}{the month in which mortality occurs}

\item{reproductive_rate}{number of spores or pest units produced by a single host under optimal weather conditions}

\item{weather}{Boolean that is true if weather is used}

\item{mortality_on}{boolean to turn host mortality on and off}

\item{short_distance_scale}{distance scale parameter for short range dispersal kernel}

\item{infected}{matrix of infected hosts}

\item{susceptible}{matrix of susceptible hosts}

\item{mortality_tracker}{matrix of 0's to track mortality per year}

\item{mortality}{matrix to track cumulative mortality}

\item{total_plants}{matrix of total plants}

\item{treatment_maps}{list of matrices where treatment or management has occurred in a given year}

\item{treatment_years}{years in which to apply treatment}

\item{temperature}{vector of matrices of temperature values used to check against lethal temperature}

\item{weather_coefficient}{vector of matrices of weather coefficients}

\item{ew_res}{east and west resolution}

\item{ns_res}{north and south resolution}

\item{time_step}{how often should spread occur}

\item{mortality_rate}{rate at which mortality occurs}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen}

\item{season_month_end}{when does spread end during the year for your pest or pathogen}

\item{start_time}{first year to start the simulation}

\item{end_time}{last year of the simulation}

\item{dispersal_kern}{what type of dispersal kernel should be used}

\item{percent_short_distance_dispersal}{what percentage of dispersal is short range versus long range}

\item{long_distance_scale}{distance scale parameter for long range dispersal kernel}

\item{treatment_method}{what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.}

\item{treatment_month}{the time during the year that treatment is applied. Currently monthly option so can be 1 -12. Default is 12.}

\item{wind_dir}{sets the wind direction}

\item{kappa}{sets the strength of the wind direction in the von-mises distribution}
}
\value{
list of vector matrices of infected and suceptible hosts per simulated year
}
\description{
PoPS c++ model
}
