% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantity_allocation_disagreement.R
\name{quantity_allocation_disagreement}
\alias{quantity_allocation_disagreement}
\title{Compares quantity and allocation disagreement of two raster datasets}
\usage{
quantity_allocation_disagreement(
  reference,
  comparison,
  use_configuration = FALSE,
  mask = NULL,
  use_distance = FALSE
)
}
\arguments{
\item{reference}{the raster with ground truth data. For all metrics expect
RMSE these values are reclassified with values > 1 becoming 1, values < 1
going to 0, and NA values staying NA.}

\item{comparison}{the raster with simulated data. For all metrics expect
RMSE these values are reclassified with values > 1 becoming 1, values < 1
going to 0.}

\item{use_configuration}{Boolean if you want to use configuration
disagreement for comparing model runs. Default is FALSE.}

\item{mask}{Used to provide a mask to remove 0's that are not true
negatives from comparisons.}

\item{use_distance}{Boolean if you want to compare distance between
simulations and observations. Default is FALSE.}
}
\value{
A data frame with spatial configuration metrics. Particularly
quantity, allocation, and total disagreement,  omission and comission, and
directional disagreement where directional disagreement.
}
\description{
Uses quantity and allocation disagreement metrics by Pontius and Millones
(2014) and omission and comission errors of comparing a modeled raster
dataset to a reference raster datatset.
}
