% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validates the accuracy of the calibrated reproductive rate and dispersal scales of the pops model.}
\usage{
validate(infected_years_file, num_iterations, number_cores = NA,
  infected_file, host_file, total_plants_file, temp = FALSE,
  temperature_coefficient_file = "", precip = FALSE,
  precipitation_coefficient_file = "", time_step = "month",
  reproductive_rate = 3, season_month_start = 1,
  season_month_end = 12, start_time = 2018, end_time = 2020,
  use_lethal_temperature = FALSE, temperature_file = "",
  lethal_temperature = -12.87, lethal_temperature_month = 1,
  mortality_on = FALSE, mortality_rate = 0, mortality_time_lag = 0,
  management = FALSE, treatment_years = c(0), treatments_file = "",
  treatment_method = "ratio", treatment_month = 12,
  percent_natural_dispersal = 1, natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy", natural_distance_scale = 21,
  anthropogenic_distance_scale = 0, natural_dir = "NONE",
  natural_kappa = 0, anthropogenic_dir = "NONE",
  anthropogenic_kappa = 0, mask = NULL, success_metric = "quantity")
}
\arguments{
\item{infected_years_file}{years of initial infection/infestation as individual locations of a pest or pathogen in raster format}

\item{num_iterations}{how many iterations do you want to run to allow the calibration to converge}

\item{number_cores}{enter how many cores you want to use (default = NA). If not set uses the # of CPU cores - 1.}

\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster file with number of hosts}

\item{total_plants_file}{path to raster file with number of total plants}

\item{temp}{allows the use of temperature coefficients to modify spread}

\item{temperature_coefficient_file}{path to raster file with temperature coefficient data for the timestep and number of years specified}

\item{precip}{allows the use of precipitation coefficients to modify spread}

\item{precipitation_coefficient_file}{path to raster file with precipitation coefficient data for the timestep and number of years specified}

\item{time_step}{how often should spread occur}

\item{reproductive_rate}{number of spores or pest units produced by a single host under optimal weather conditions}

\item{season_month_start}{when does spread first start occurring in the year for your pest or pathogen}

\item{season_month_end}{when does spread end during the year for your pest or pathogen}

\item{start_time}{first year to start the simulation}

\item{end_time}{last year of the simulation}

\item{use_lethal_temperature}{does your pest or pathogen have a temperature at which it cannot survive}

\item{temperature_file}{path to raster file with temperature data for minimum temperature}

\item{lethal_temperature}{the temperature at which lethal temperature related mortality occurs for your pest or pathogen}

\item{lethal_temperature_month}{the month in which lethal temperature related mortality occurs for your pest or pathogen}

\item{mortality_on}{boolean to turn host mortality on and off}

\item{mortality_rate}{rate at which mortality occurs}

\item{mortality_time_lag}{time lag from infection until mortality can occur in years}

\item{management}{boolean to allow use of managemnet}

\item{treatment_years}{years in which to apply treatment}

\item{treatments_file}{path to raster file with treatment data by years}

\item{treatment_method}{what method to use when applying treatment one of ("ratio" or "all infected"). ratio removes a portion of all infected and susceptibles, all infected removes all infected a portion of susceptibles.}

\item{treatment_month}{the time during the year that treatment is applied. Currently monthly option so can be 1 -12. Default is 12.}

\item{percent_natural_dispersal}{what percentage of dispersal is natural range versus anthropogenic range}

\item{natural_kernel_type}{what type of dispersal kernel should be used for natural dispersal}

\item{anthropogenic_kernel_type}{what type of dispersal kernel should be used for anthropogenic dispersal}

\item{natural_distance_scale}{distance scale parameter for natural range dispersal kernel}

\item{anthropogenic_distance_scale}{distance scale parameter for anthropogenic range dispersal kernel}

\item{natural_dir}{sets the predominate direction of natural dispersal usually due to wind}

\item{natural_kappa}{sets the strength of the natural direction in the von-mises distribution}

\item{anthropogenic_dir}{sets the predominate direction of anthropogenic dispersal usually due to human movement typically over long distances (e.g. nursery trade, movement of firewood, etc..)}

\item{anthropogenic_kappa}{sets the strength of the anthropogenic direction in the von-mises distribution}
}
\value{
a dataframe of the variables saved and their success metrics for each run
}
\description{
This function uses the quantity, allocation, and configuration disagreement to validate the model
across the landscape using the parameters from the calibrate function. Ideally the model is calibrated
with 2 or more years of data and validated for the last year or if you have 6 or more years of data 
then the model can be validated for the final 2 years.
}
\examples{
infected_years_file <- system.file("extdata", "SODexample", "initial_infection2005.tif", 
package = "PoPS")
num_iterations <- 100
number_cores <- NA

infected_file <- system.file("extdata", "SODexample", "initial_infection2004.tif", 
package = "PoPS")
host_file <- system.file("extdata", "SODexample", "host.tif", package = "PoPS")
total_plants_file <- system.file("extdata", "SODexample", "all_plants.tif", package = "PoPS")
temperature_coefficient_file <- system.file("extdata", "SODexample", "weather.tif", 
package = "PoPS")
treatments_file <- system.file("extdata", "SODexample", "management2005.tif", package = "PoPS")

params <- validate(infected_years_file, num_iterations, number_cores,
infected_file, host_file, total_plants_file, reproductive_rate = 1.0,
use_lethal_temperature = FALSE, temp = TRUE, precip = FALSE, management = TRUE, 
mortality_on = TRUE, temperature_file = "", temperature_coefficient_file, 
precipitation_coefficient_file ="", treatments_file,
season_month_start = 1, season_month_end = 12, time_step = "month",
start_time = 2005, end_time = 2005, treatment_years = c(2005),
dispersal_kern = "cauchy", percent_short_distance_dispersal = 1.0,
short_distance_scale = 20.57, long_distance_scale = 0.0,
lethal_temperature = -12.87, lethal_temperature_month = 1,
mortality_rate = 0.05, mortality_time_lag = 2,
wind_dir = "NONE", kappa = 0)

}
