% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validates the accuracy of the calibrated reproductive rate and dispersal
scales of the pops model.}
\usage{
validate(
  infected_years_file,
  number_of_iterations,
  number_of_cores = NA,
  parameter_means,
  parameter_cov_matrix,
  infected_file,
  host_file,
  total_populations_file,
  temp = FALSE,
  temperature_coefficient_file = "",
  precip = FALSE,
  precipitation_coefficient_file = "",
  model_type = "SI",
  latency_period = 0,
  time_step = "month",
  season_month_start = 1,
  season_month_end = 12,
  start_date = "2008-01-01",
  end_date = "2008-12-31",
  use_lethal_temperature = FALSE,
  temperature_file = "",
  lethal_temperature = -12.87,
  lethal_temperature_month = 1,
  mortality_on = FALSE,
  mortality_rate = 0,
  mortality_time_lag = 0,
  management = FALSE,
  treatment_dates = c(""),
  treatments_file = "",
  treatment_method = "ratio",
  natural_kernel_type = "cauchy",
  anthropogenic_kernel_type = "cauchy",
  natural_dir = "NONE",
  anthropogenic_dir = "NONE",
  pesticide_duration = 0,
  pesticide_efficacy = 1,
  mask = NULL,
  success_metric = "quantity",
  output_frequency = "year",
  output_frequency_n = 1,
  movements_file = "",
  use_movements = FALSE,
  start_exposed = FALSE,
  generate_stochasticity = TRUE,
  establishment_stochasticity = TRUE,
  movement_stochasticity = TRUE,
  deterministic = FALSE,
  establishment_probability = 0.5,
  dispersal_percentage = 0.99,
  quarantine_areas_file = "",
  use_quarantine = FALSE,
  use_spreadrates = FALSE,
  use_overpopulation_movements = FALSE,
  overpopulation_percentage = 0,
  leaving_percentage = 0,
  leaving_scale_coefficient = 1,
  exposed_file = "",
  write_outputs = "None",
  output_folder_path = ""
)
}
\arguments{
\item{infected_years_file}{years of initial infection/infestation as
individual locations of a pest or pathogen in raster format}

\item{number_of_iterations}{how many iterations do you want to run to allow
the calibration to converge at least 10}

\item{number_of_cores}{enter how many cores you want to use (default = NA).
If not set uses the # of CPU cores - 1. must be an integer >= 1}

\item{parameter_means}{the parameter means from the abc calibration function
(posterior means)}

\item{parameter_cov_matrix}{the parameter covariance matrix from the abc
calibration function (posterior covairance matrix)}

\item{infected_file}{path to raster file with initial infections}

\item{host_file}{path to raster files with number of hosts and standard
deviation on those estimates can be based in 3 formats (a single file with
number of hosts, a single file with 2 layers number of hosts and standard
deviation, or two files 1 with number of hosts and the other with standard
deviation of those estimates)}

\item{total_populations_file}{path to raster file with number of total
populations of all hosts and non-hosts (must be only hosts if using movement)}

\item{temp}{boolean that allows the use of temperature coefficients to
modify spread (TRUE or FALSE)}

\item{temperature_coefficient_file}{path to raster file with temperature
coefficient data for the timestep and number of years specified}

\item{precip}{boolean that allows the use of precipitation coefficients to
modify spread (TRUE or FALSE)}

\item{precipitation_coefficient_file}{path to raster file with precipitation
coefficient data for the timestep and number of years specified}

\item{model_type}{What type of model most represents your sysetm. Options
are "SEI" (Susceptible - Exposed - Infected/Infested) or "SI"
(Susceptible - Infected/Infested). Default value is "SI".}

\item{latency_period}{How many times steps does it take to for exposed
populations become infected/infested. This is an integer value and must be
greater than 0 if model type is SEI.}

\item{time_step}{how often should spread occur options: ('day', 'week',
'month')}

\item{season_month_start}{when does spread first start occurring in the year
for your pest or pathogen (integer value between 1 and 12)}

\item{season_month_end}{when does spread end during the year for your pest
or pathogen (integer value between 1 and 12)}

\item{start_date}{date to start the simulation with format ('YYYY_MM_DD')}

\item{end_date}{date to end the simulation with format ('YYYY_MM_DD')}

\item{use_lethal_temperature}{a boolean to answer the question: does your
pest or pathogen have a temperature at which it cannot survive?
(TRUE or FALSE)}

\item{temperature_file}{path to raster file with temperature data for
minimum temperature}

\item{lethal_temperature}{the temperature in degrees C at which lethal
temperature related mortality occurs for your pest or pathogen (-50 and 60)}

\item{lethal_temperature_month}{the month in which lethal temperature
related mortality occurs for your pest or pathogen integer value between 1
and 12}

\item{mortality_on}{boolean to turn host mortality on and off
(TRUE or FALSE)}

\item{mortality_rate}{rate at which mortality occurs value between 0 and 1}

\item{mortality_time_lag}{time lag from infection until mortality can occur
in time steps integer >= 1}

\item{management}{boolean to allow use of management (TRUE or FALSE)}

\item{treatment_dates}{dates in which to apply treatment list with format
('YYYY_MM_DD') (needs to be the same length as treatment_file and
pesticide_duration)}

\item{treatments_file}{path to raster files with treatment data by dates.
Needs to be a list of files the same length as treatment_dates and
pesticide_duration.}

\item{treatment_method}{what method to use when applying treatment one of
("ratio" or "all infected"). ratio removes a portion of all infected and
susceptibles, all infected removes all infected a portion of susceptibles.}

\item{natural_kernel_type}{what type of dispersal kernel should be used for
natural dispersal. Current dispersal kernel options are
 ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law',
 'hyperbolic secant', 'gamma', 'weibull',
 'logistic')}

\item{anthropogenic_kernel_type}{what type of dispersal kernel should be
used for anthropogenic dispersal. Current dispersal kernel options are
 ('cauchy', 'exponential', 'uniform', 'deterministic neighbor','power law',
 'hyperbolic secant', 'gamma', 'weibull',
 'logistic')}

\item{natural_dir}{sets the predominate direction of natural dispersal
usually due to wind values ('N', 'NW', 'W', 'SW', 'S', 'SE', 'E', 'NE',
'NONE')}

\item{anthropogenic_dir}{sets the predominate direction of anthropogenic
dispersal usually due to human movement typically over long distances (e.g.
nursery trade, movement of firewood, etc..) ('N', 'NW', 'W', 'SW', 'S',
'SE', 'E', 'NE', 'NONE')}

\item{pesticide_duration}{how long does the pestcide (herbicide, vaccine,
etc..) last before the host is susceptible again. If value is 0 treatment
is a culling (i.e. host removal) not a pesticide treatment. (needs to be the
 same length as treatment_dates and treatment_file)}

\item{pesticide_efficacy}{how effictive is the pesticide at preventing the
disease or killing the pest (if this is 0.70 then when applied it
successfully treats 70 percent of the plants or animals)}

\item{mask}{Raster file used to provide a mask to remove 0's that are not
true negatives from comparisons (e.g. mask out lakes and oceans from statics
if modeling terrestrial species).}

\item{success_metric}{Choose which success metric to use for calibration.
Choices are "quantity", "quantity and configuration", "residual error" and
"odds ratio". Default is "quantity"}

\item{output_frequency}{sets when outputs occur either ('year', 'month',
'week', 'day' or 'time step')}

\item{output_frequency_n}{sets number units from output_frequency in which to
export model results.}

\item{movements_file}{this is a csv file with columns lon_from, lat_from,
lon_to, lat_to, number of animals, and date.}

\item{use_movements}{this is a boolean to turn on use of the movement module.}

\item{start_exposed}{Do your initial conditions start as exposed or infected
(only used if model_type is "SEI"). Default False. If this is TRUE need to
have both an infected_file (this can be a raster of all 0's) and exposed_file}

\item{generate_stochasticity}{Boolean to indicate whether to use
stochasticity in reproductive functions default is TRUE}

\item{establishment_stochasticity}{Boolean to indicate whether to use
stochasticity in establishment functions default is TRUE}

\item{movement_stochasticity}{Boolean to indicate whether to use
stochasticity in movement functions default is TRUE}

\item{deterministic}{Boolean to indicate whether to use a deterministic
dispersal kernel default is FALSE}

\item{establishment_probability}{Threshold to determine establishment if
establishment_stochasticity is FALSE (range 0 to 1, default = 0.5)}

\item{dispersal_percentage}{Percentage of dispersal used to calculate the
bounding box for deterministic dispersal}

\item{quarantine_areas_file}{path to raster file with quarantine boundaries
used in calculating likelihood of quarantine escape if use_quarantine is
TRUE}

\item{use_quarantine}{boolean to indicate whether or not there is a
quarantine area if TRUE must pass in a raster file indicating the
quarantine areas (default = FALSE)}

\item{use_spreadrates}{boolean to indicate whether or not to calculate
spread rates}

\item{use_overpopulation_movements}{boolean to indicate whether to use
the overpopulation pest movement module (driven by the natural kernel with
its scale parameter modified by a coefficient)}

\item{overpopulation_percentage}{percentage of occupied hosts when the cell
is considered to be overpopulated}

\item{leaving_percentage}{percentage of pests leaving an overpopulated cell}

\item{leaving_scale_coefficient}{coefficient to multiply scale parameter of
the natural kernel (if applicable)}

\item{exposed_file}{a file with the exposed for the current}

\item{write_outputs}{Either c("summary outputs", or "None"). If not
"None" output folder path must be provided.}

\item{output_folder_path}{this is the full path with either / or \\ (e.g.,
"C:/user_name/desktop/pops_sod_2020_2023/outputs/")}
}
\value{
a dataframe of the variables saved and their success metrics for
each run
}
\description{
This function uses the quantity, allocation, and configuration disagreement
to validate the model across the landscape using the parameters from the
calibrate function. Ideally the model is calibrated with 2 or more years of
data and validated for the last year or if you have 6 or more years of data
then the model can be validated for the final 2 years.
}
